requirements.txt
Django==3.2
pycryptodome
pymysql
eciespy
